test fail get cities
